+++ src/App.vue
@@ -194,3 +194,7 @@ export default {
     },
 }
 </script>
+
+<style>
+
+</style>